backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
httpx
python-dotenv
fitz